Loose ToggleSetter implementation
ToggleSetter is the new way of doing things! Instead of creating a component for every Setting, there will be a Component for every type of Selectable instead. The setting can then be picked from a list of all available settings that are compatible. This list is extendible by other developers of course by simply creating a new setting.