Editor Implementation
The workflow is falling into place, as things will be handled quite differently from now on. There will only be components for every type of Selectable. Then, these Components hold a reference to their settings that implement their own Update behaviour.

The implementation of a setting is done via an interface. What's great about that is that we can create a Setting that implements behaviour both for a dropdown and a slider for example.

The last thing to note is that settings you want to show up in the inspector will need an attribute of UISetting or else it will be ommited.